requirements.txt
flask~=2.3.2
flask-smorest
sqlalchemy
flask-sqlalchemy
python-dotenv
marshmallow~=3.20.1
psycopg2